+++ tests/test_generalutils.py
@@ -11,12 +11,12 @@ from generalutils.constants import HttpStatusCode
 
 class TestGeneralUtils(unittest.TestCase):
     
-    def test_DirectoryExists(self):
+    def test_DirectoryExistsException(self):
         '''Test whether a certain directory exists'''
 
         self.assertTrue(guard.Filesystem.PathExist(os.getcwd()))
 
-    def test_DirectoryDoesNotExists(self):
+    def test_DirectoryDoesNotExistsException(self):
         '''Negative test to check whether a certain directory doesn't exists'''
         
         # Random not existing folder path
@@ -30,13 +30,26 @@ class TestGeneralUtils(unittest.TestCase):
             guard.Filesystem.PathExist(path)
 
         self.assertTrue(exception in str(context.exception))
+
+    def test_DirectoryExists(self):
+        '''Test whether a certain directory exists'''
+
+        self.assertTrue(guard.Filesystem.IsPath(os.getcwd()))
+
+    def test_DirectoryDoesNotExists(self):
+        '''Negative test to check whether a certain directory doesn't exists'''
+        
+        # Random not existing folder path
+        path = f'D:/{random.randint(100,10000)}'
+
+        self.assertFalse(guard.Filesystem.IsPath(path))
     
     def test_ObjectIsNotNoneOrEmpty(self):
         '''Test whether object is not none or empty'''
 
         word = "Not empty"
 
-        self.assertTrue(guard.Collections.IsNotNoneOrEmpty(word))
+        self.assertTrue(guard.Collections.NotNoneOrEmpty(word))
 
     def test_ObjectIsEmpty(self):
         '''Negative test to check whether a object is empty'''
@@ -47,7 +60,7 @@ class TestGeneralUtils(unittest.TestCase):
         exception = f"Object - '{word}' is none or empty"
 
         with self.assertRaises(Exception) as context:
-            guard.Collections.IsNotNoneOrEmpty(word)
+            guard.Collections.NotNoneOrEmpty(word)
 
         self.assertTrue(exception in str(context.exception))
 
@@ -60,7 +73,7 @@ class TestGeneralUtils(unittest.TestCase):
         exception = f"Object - '{word}' is none or empty"
 
         with self.assertRaises(Exception) as context:
-            guard.Collections.IsNotNoneOrEmpty(word)
+            guard.Collections.NotNoneOrEmpty(word)
 
         self.assertTrue(exception in str(context.exception))
     
@@ -105,14 +118,14 @@ class TestGeneralUtils(unittest.TestCase):
 
         arg = ["this", "is", "a", "test"]
 
-        self.assertTrue(guard.Argument.Valid(arg))
+        self.assertTrue(guard.Argument.IsValid(arg))
     
     def test_InvalidArgument(self):
         '''Test whether an argument is invalid'''
 
         arg = None
 
-        self.assertFalse(guard.Argument.Valid(arg))
+        self.assertFalse(guard.Argument.IsValid(arg))
 
 if __name__ == '__main__':
 	unittest.main()